requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
Babel==2.6.0
bleach==4.1.0
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.14
django-allauth==0.42.0
django-anymail==8.6
django-autocomplete-light==3.2.10
django-cors-headers==3.10.1
django-date-extensions==3.1.2
django-filter==2.4.0
django-guardian==2.4.0
django-modeltranslation==0.18.2
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==0.0.1
django-rest-knox==3.1.5
django-rest-swagger==2.2.0
django-summernote==0.8.20.0
django-user-agents==0.3.2
django-utils-six==2.0
django-webpack-loader==0.6.0
djangorestframework==3.13.1
docopt==0.4.0
drf-yasg==1.21.4
gunicorn==20.1.0
html5lib==1.0.1
httpie==0.9.9
idna==2.6
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.1.0
Jinja2==2.10
mailchimp==2.0.9
MarkupSafe==1.0
nodeenv==1.3.0
oauthlib==2.1.0
openapi-codec==1.3.2
packaging==21.3
pdfrw==0.4
Pillow==5.1.0
psycopg2-binary==2.9.5
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pyOpenSSL==18.0.0
pyotp==2.2.6
pyparsing==3.0.9
PyPDF2==1.26.0
Pyphen==0.9.5
PySocks==1.6.8
python-dateutil==2.8.1
python-dotenv==0.20.0
python-memcached==1.59
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2022.4
PyYAML==6.0
qrcode==6.0
qrplatba==0.3.4
reportlab==3.5.9
requests==2.20.0
requests-oauthlib==0.8.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.15.0
six==1.11.0
social-auth-core==4.0.2
sqlparse==0.4.2
stripe==2.17.0
tinycss2==0.6.1
twilio==6.15.2
typing_extensions==4.1.1
ua-parser==0.15.0
uritemplate==3.0.0
urllib3==1.22
user-agents==2.2.0
WeasyPrint==0.42.3
webencodings==0.5.1
zipp==3.6.0
